+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added 
-- ANG file writer for CrystalMap objects (via orix.io.save())
+- .ang file writer for CrystalMap objects (via orix.io.save())
 - Overloaded division for Vector3d (left hand side) by numbers and suitably shaped array-like objects
 
 ### Changed
@@ -20,5 +20,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - CrystalMap properties allow arrays with number of dimensions greater than 2
-- ANG file reader now recognises phase IDs defined in the header
+- .ang file reader now recognises phase IDs defined in the header
 - EMsoft file reader reads unrefined Euler angles correctly